Created a separate file for the Vacancy class
